TVG-102 Blueprint for the SearchItem endpoints
