+++ api_blueprints/__init__.py
@@ -1,0 +1,2 @@
+from api_blueprints.search_items import search_items_blueprint
+from api_blueprints.shows import shows_blueprint
